NBD server: handle NBD_CMD_DISC correctly instead of returning EINVAL
We simply return unit in case of the Disc command, since the user of the
serve function should close the TCP session. Closing the session is
required for the NBD server implementation to be correct, because the
NBD protocol requires the server to do this after a disconnect request
is received.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>