Add support for NBD_OPT_GO
Now qemu-img can connect to this NBD server using NBD_OPT_GO, instead of
listing the export names using NBD_OPT_LIST and then connecting with
NBD_OPT_EXPORT_NAME.

Since NBD_OPT_INFO is not implemented, which is closely related to
NBD_OPT_GO, this might be confusing for some clients. I have only
implemented the bare minimum required by the NBD protocol and to get
qemu-img 2.11.1 to work. For example, we currently ignore all the
information requests in NBD_OPT_GO, and only send the required
NBD_INFO_EXPORT.

Currently I've implemented NBD_OPT_GO in a way that the current
interface does not change - we send the last option haggling response
from Server.connect, both for NBD_OPT_EXPORT_NAME and NBD_OPT_GO. This
is because we need the Mirage block device to get some info about the
export. But we might have to change things in the future, if we want to
support NBD_OPT_INFO somehow.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>